Print possible missed attack spells. WATCH FOR THESE
I was setting attacking when I called the attack. If I were cc'd during
that or otherwise unable to attack I'd have to wait a full attack cycle.
If I didn't it was causing odd behavior. I think that's caused by not
having all of the attack spells which resulted in not triggering a
windup and therefor clip. So instead when I call AA it will print the
next spell name if it isn't an already recorded attack.